fix(cms): use hooks instead of jobs for newsletter
